requirements/req.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
fastapi==0.104.1
numpy==1.26.2
pandas==2.1.3
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
SQLAlchemy==2.0.23
uvicorn==0.24.0.post1
